Slow down by a factor of 3 to compensate for unknown performance boost.
Whoa, somehow my code sped up by a factor of 3.  I can only assume
some dramatic improvement in the compiler or in one of the libraries I
depend on.  This highlights the problem with the Arduino approach,
where it always builds from source when you flash the chip, rather
than producing a binary you can save and reuse.

Slowing down the color change by a factor of 3 brings it approximately
back to where it was.